+++ tests/testthat/test_normalize.R
@@ -297,7 +297,11 @@ test_that("fscale", {
   expect_lt(abs(integrate_spct(fscale(my.spct, f = "total")) - 1), 1e6)
   expect_lt(abs(average_spct(fscale(my.spct, f = "mean")) - 1), 1e6)
   expect_warning(irrad(fscale(my.spct, f = "mean")))
-  expect_false(suppressWarnings(is.na(irrad(fscale(my.spct, f = "mean")))))
+  #  behaviour changed in 0.13.3
+  #  expect_false(suppressWarnings(is.na(irrad(fscale(my.spct, f = "mean")))))
+  expect_true(suppressWarnings(is.na(irrad(fscale(my.spct, f = "mean")))))
+  expect_no_warning(irrad(fscale(my.spct, f = "mean"), allow.scaled = TRUE))
+  expect_false(is.na(irrad(fscale(my.spct, f = "mean"), allow.scaled = TRUE)))
   expect_named(fscale(my.spct), names(my.spct))
   expect_equal(class(fscale(my.spct)), class(my.spct))
   expect_error(fscale(my.spct, range = 281))